Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Well the need for this change came when trying to run the "su" binary in the SDK build for Honeycomb. _mktemp does not seem to be present in the HC bionic libs. Furthermore, just using mktemp instead does not break anything on the gingerbread build.

Patch-set: 1