Update patch set 1
Patch Set 1:

Sean - It's possible to determine if this is working using an instrumented kernel that replaces everything in mmc_erase() with a printk() combined with returning -EOPNOTSUPP all the time.  David linked to the commits for one such test I ran to prove that Samsung's own stock recoveries are dangerous.

I'll work on running a test with an instrumented kernel tomorrow.

Patch-set: 1