Update patch set 1
Patch Set 1:

I patched this with LZ4 support but haven't finished the kernel side quite yet. Will try to get that done today.

What I'm noticing is that on some devices when the LMK kicks in, the system can thrash a bit when launching new apps if you go overboard with compressing stuff. LZ4 might help with this since it's so insanely fast, the downside is that it doesn't compress as well as gzip.

Patch-set: 1