Update patch set 1
Patch Set 1: Code-Review+1

The glibc bug claims that malloc may set errno even when returning a valid ptr.

Questions for my own edification...

Isn't that in itself a bug?  Why are they working around it instead of fixing it?

Why and under what circumstances does this bug manifest?  How would we be assured that the second scandir would succeed?

What versions of glibc are affected?  Can we correlate these with particular distros and avoid them?

Nevertheless, I see this was picked from AOSP master, so it should probably go in anyway.

Patch-set: 1