Update patch set 1
Patch Set 1:

Not aware fastboot is still linking with  f2fs_ioutils and f2fs_dlutils libraries.
Patch summited to link only f2fs_dlutils.

The reason for the change here is that there is duplicated symbol in make_f2fs and the dynamic loaded library (e.g. f2fs_configuration config and f2fs_sparse_file). In Linux the symbol in make_f2fs will override the symbol (or have a priority) to the symbol in the dynamic loaded library but in darwin it is not.

Patch-set: 1