Replace lazy_static with once_cell
Using lazy_static is now discouraged as unmaintained
and once_cell is the recommended replacement.

On top of that a similar implementation found in
once_cell is being tracked for inclusion under the
`lazy_cell` feature gate [0]

Also there's no need to lazy init the Timestamp
Context as it's const.

[0] https://github.com/rust-lang/rust/issues/109736

Test: m libprofcollectd
Change-Id: I6314a47519b87e027005ccb6173a742a18050305