tools/check_elf_alignment.sh: Add zip alignment check
If an APK is passed in, then check the zipalignment as well to see if
it's 16K aligned. This will help cover both requirements -- zipalignment
and ELF alignment.

Also, add a print message after checking ELF alignment to indicate that
it was successful. This helps indicate success when there are no shared
libraries in the APK.

Bug: 337895632
Change-Id: I2578dba77e13570aa5b880f950058cfcd939857f