Automatically delete trace files after generating report
Summary: After switching to a time-based approach for report generation, we want to
ensure that we don't accidentally include repeated trace files. Thus, after
creating a zip file for the report, we delete the originals.

Test: Ran adb shell device_config put profcollect_native_boot enabled true
and rebooted the device to enable profcollectd.
Ranprofcollectctl once to collect trace files, profcollectctl process to
process them, and then profcollectctl report to generate the report.
Verified that the processed trace files were removed after profcollectctl report
was run & that the config.json file remained.

Tags:
Change-Id: Ic6ab934f52c7af277817ce992faebd96fcde223f
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>